# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lexiflow API
  description: |
    REST API for the LingQ Clone language-learning app: authentication, library/courses,
    lesson reader, vocabulary, phrases, uploads, and optional LingQ integration.

    **Authentication:** Most routes require a JWT in the `Authorization` header as `Bearer <token>`.
    Obtain a token from `POST /auth/login` or `POST /auth/register` (register does not return a token; log in after sign-up).
  version: 1.0.0
  license:
    name: ISC
    identifier: ISC
servers:
  - url: http://localhost:3000/api/v1
    description: Local or Docker (default port 3000)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /auth/login (`token` field)
...
paths:
...
  /library/courses:
    get:
      tags: [Library]
      summary: List courses owned by the current user
      operationId: listMyCourses
      security:
        - bearerAuth: []
      parameters:
        - name: lang
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Course rows
        '401':
          description: Unauthorized
    post:
      tags: [Library]
      summary: Create a course
      operationId: createCourse
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                languageCode:
                  type: string
                level:
                  type: string
                imageUrl:
                  type: string
                isPublic:
                  type: boolean
      responses:
        '200':
          description: Created course
        '401':
          description: Unauthorized
  /library/courses/{id}:
    delete:
      tags: [Library]
      summary: Delete a course (optionally require confirm for courses with lessons)
      operationId: deleteCourse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: confirm
          in: query
          description: Set to `true` to delete a course that still has lessons
          schema:
            type: string
            enum: ['true']
      responses:
        '200':
          description: Deleted
        '400':
          description: confirm_required with lesson count
        '404':
          description: Not found or not owner
  /library/courses/{id}/lessons:
    get:
      tags: [Library]
      summary: Lessons in a course with progress and course-level stats
      operationId: getCourseLessons
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object with `course` and `lessons`
        '404':
          description: Course not found
  /library/feed/{langCode}:
    get:
      tags: [Library]
      summary: Lesson feed (excludes courses user is enrolled in)
      operationId: getLessonFeed
      security:
        - bearerAuth: []
      parameters:
        - name: langCode
          in: path
          required: true
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Feed items with progress fields
        '400':
          description: Invalid language
  /library/my-lessons:
    get:
      tags: [Library]
      summary: Lessons from enrolled courses
      operationId: getMyLessons
      security:
        - bearerAuth: []
      parameters:
        - name: lang
          in: query
          required: true
          schema:
            type: string
        - name: completed
          in: query
          description: Filter by `"true"` or `"false"`
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lesson list
  /library/check-completions:
    post:
      tags: [Library]
      summary: Batch completion check (currently disabled in implementation)
      operationId: checkCompletions
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                lang:
                  type: string
      responses:
        '200':
          description: Result with updatedCount
  /library/continue-studying:
    get:
      tags: [Library]
      summary: Recently read lessons (up to 4)
      operationId: continueStudying
      security:
        - bearerAuth: []
      parameters:
        - name: lang
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recent lessons
  /library/guided-courses:
    get:
      tags: [Library]
      summary: Courses with enrollment and progress aggregates
      operationId: getGuidedCourses
      security:
        - bearerAuth: []
      parameters:
        - name: lang
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Course stats array
  /library/lessons/{id}/bookmark:
    post:
      tags: [Library]
      summary: Toggle lesson bookmark
      operationId: toggleLessonBookmark
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'JSON body with a `bookmarked` boolean property.'
  /library/lingq-import:
    post:
      tags: [Library]
      summary: One-time import from LingQ API (per user language)
      operationId: lingqImport
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                apiKey:
                  type: string
                languageCode:
                  type: string
                courseCount:
                  type: integer
                lessonsPerCourse:
                  type: integer
      responses:
        '200':
          description: Import result
        '403':
          description: Already imported for this language
        '500':
          description: Import failed
...